python/audio-summarizer/simple/requirements.txt
python-dotenv==1.1.1
ibm-watson==10.0.0
ibm_watsonx_ai==1.3.32